cuongnb fix invoice
