+++ db.json
@@ -1,343 +1,5 @@
 {
   "invoice": [
-    {
-<<<<<<< HEAD
-      "id": "14",
-      "reservation_id": "14",
-      "totalAmount": 145000,
-      "invoice_item": [
-        {
-          "id": 5,
-          "name": "Phở gà",
-          "quantity": 3,
-          "price": 40000,
-          "total": 120000
-        },
-        {
-          "id": 39,
-          "name": "undefined",
-          "quantity": 1,
-          "price": 25000,
-          "total": 25000
-        }
-      ]
-    },
-    {
-      "id": "14",
-      "reservation_id": "14",
-      "totalAmount": 120000,
-      "invoice_item": [
-        {
-          "id": 5,
-          "name": "Phở gà",
-          "quantity": 3,
-          "price": 40000,
-          "total": 120000
-        }
-      ]
-    },
-    {
-      "id": "13",
-      "reservation_id": "13",
-      "totalAmount": 45000,
-      "invoice_item": [
-        {
-          "id": 6,
-          "name": "Bún thịt nướng",
-          "quantity": 1,
-          "price": 45000,
-          "total": 45000
-        }
-      ]
-    },
-    {
-      "id": "14",
-      "reservation_id": "14",
-      "totalAmount": 120000,
-      "invoice_item": [
-        {
-          "id": 5,
-          "name": "Phở gà",
-          "quantity": 3,
-          "price": 40000,
-          "total": 120000
-        }
-      ]
-    },
-    {
-      "id": "13",
-      "reservation_id": "13",
-      "totalAmount": 45000,
-      "invoice_item": [
-        {
-          "id": 6,
-          "name": "Bún thịt nướng",
-          "quantity": 1,
-          "price": 45000,
-          "total": 45000
-        }
-      ]
-    },
-    {
-      "id": "12",
-      "reservation_id": "12",
-      "totalAmount": 40000,
-      "invoice_item": [
-        {
-          "id": 40,
-          "name": "Cháo thịt bằm",
-          "quantity": 1,
-          "price": 35000,
-          "total": 35000
-        },
-=======
-      "id": "1",
-      "reservation_id": "1",
-      "totalAmount": 145000,
-      "list_table": [
->>>>>>> 5e1202ec1c32a56050cf6389f34f45df7d6eeadb
-        {
-          "id": 1,
-          "name": "Bàn 1"
-        },
-        {
-          "id": 2,
-          "name": "Bàn 2"
-        }
-      ],
-      "invoice_item": [
-        {
-          "id": 6,
-          "name": "Bún thịt nướng",
-          "quantity": 1,
-          "price": 45000,
-          "total": 45000
-        },
-        {
-          "id": 4,
-          "name": "Cháo thịt bằm",
-          "quantity": 1,
-          "price": 35000,
-          "total": 35000
-        },
-        {
-          "id": 5,
-          "name": "Phở gà",
-          "quantity": 1,
-          "price": 40000,
-          "total": 40000
-        },
-        {
-          "id": 3,
-          "name": "Xôi mặn",
-          "quantity": 1,
-          "price": 25000,
-          "total": 25000
-        }
-      ]
-    },
-    {
-      "id": "1",
-      "reservation_id": "1",
-      "totalAmount": 50000,
-      "invoice_item": [
-        {
-          "id": 1,
-          "name": "undefined",
-          "quantity": 1,
-          "price": 20000,
-          "total": 20000
-        },
-        {
-          "id": 2,
-          "name": "undefined",
-          "quantity": 1,
-          "price": 30000,
-          "total": 30000
-        }
-      ]
-    },
-    {
-      "id": "13",
-      "reservation_id": "13",
-      "totalAmount": 110000,
-      "invoice_item": [
-        {
-          "id": 37,
-          "name": "Bánh mì trứng",
-          "quantity": 1,
-          "price": 20000,
-          "total": 20000
-        },
-        {
-          "id": 38,
-          "name": "Bánh cuốn",
-          "quantity": 1,
-          "price": 30000,
-          "total": 30000
-        },
-        {
-          "id": 39,
-          "name": "Xôi mặn",
-          "quantity": 1,
-          "price": 25000,
-          "total": 25000
-        },
-        {
-          "id": 40,
-          "name": "Cháo thịt bằm",
-          "quantity": 1,
-          "price": 35000,
-          "total": 35000
-        }
-      ]
-    },
-    {
-      "id": "13",
-      "reservation_id": "13",
-      "totalAmount": 75000,
-      "invoice_item": [
-        {
-          "id": 37,
-          "name": "Bánh mì trứng",
-          "quantity": 1,
-          "price": 20000,
-          "total": 20000
-        },
-        {
-          "id": 38,
-          "name": "Bánh cuốn",
-          "quantity": 1,
-          "price": 30000,
-          "total": 30000
-        },
-        {
-          "id": 39,
-          "name": "Xôi mặn",
-          "quantity": 1,
-          "price": 25000,
-          "total": 25000
-        }
-      ]
-    },
-    {
-      "id": "14",
-      "reservation_id": "14",
-      "totalAmount": 225000,
-      "invoice_item": [
-        {
-          "id": 37,
-          "name": "Bánh mì trứng",
-          "quantity": 2,
-          "price": 20000,
-          "total": 40000
-        },
-        {
-          "id": 46,
-          "name": "Bánh mì chả cá",
-          "quantity": 1,
-          "price": 22000,
-          "total": 22000
-        },
-        {
-          "id": 57,
-          "name": "Thịt kho tàu",
-          "quantity": 1,
-          "price": 70000,
-          "total": 70000
-        },
-        {
-          "id": 56,
-          "name": "Cơm bò nướng",
-          "quantity": 1,
-          "price": 63000,
-          "total": 63000
-        },
-        {
-          "id": 38,
-          "name": "Bánh cuốn",
-          "quantity": 1,
-          "price": 30000,
-          "total": 30000
-        }
-      ]
-    },
-    {
-      "id": "14",
-      "reservation_id": "14",
-      "totalAmount": 195000,
-      "invoice_item": [
-        {
-          "id": 37,
-          "name": "Bánh mì trứng",
-          "quantity": 2,
-          "price": 20000,
-          "total": 40000
-        },
-        {
-          "id": 46,
-          "name": "Bánh mì chả cá",
-          "quantity": 1,
-          "price": 22000,
-          "total": 22000
-        },
-        {
-          "id": 57,
-          "name": "Thịt kho tàu",
-          "quantity": 1,
-          "price": 70000,
-          "total": 70000
-        },
-        {
-          "id": 56,
-          "name": "Cơm bò nướng",
-          "quantity": 1,
-          "price": 63000,
-          "total": 63000
-        }
-      ]
-    },
-    {
-      "id": "1",
-      "reservation_id": "1",
-      "totalAmount": 159000,
-      "invoice_item": [
-        {
-          "id": 1,
-          "name": "undefined",
-          "quantity": 1,
-          "price": 20000,
-          "total": 20000
-        },
-        {
-          "id": 2,
-          "name": "undefined",
-          "quantity": 1,
-          "price": 30000,
-          "total": 30000
-        },
-        {
-          "id": 10,
-          "name": "undefined",
-          "quantity": 1,
-          "price": 22000,
-          "total": 22000
-        },
-        {
-          "id": 9,
-          "name": "undefined",
-          "quantity": 1,
-          "price": 32000,
-          "total": 32000
-        },
-        {
-          "id": 11,
-          "name": "undefined",
-          "quantity": 1,
-          "price": 55000,
-          "total": 55000
-        }
-      ]
-    }
+    
   ]
 }
